Move `main()` and its static methods into `Simulation`
